requirements.txt
fastapi~=0.115.11
uvicorn==0.23.2
python-jose~=3.4.0
passlib==1.7.4
pydantic~=2.11.3
python-multipart==0.0.6
python-dotenv==1.0.0
email-validator==2.0.0
pymongo==4.2.0
motor==3.1.1
bcrypt
httpx~=0.28.1
jose~=1.0.0